requirements.txt
# Notebook
jupyter
ipympl
# Data science
pandas
scikit-learn
scikit-optimize
matplotlib
# Track experiments
wandb
# TF
#tensorflow==2.7.0
# tf 2.7 cpu:
https://storage.googleapis.com/tensorflow/linux/cpu/tensorflow_cpu-2.7.0-cp37-cp37m-manylinux2010_x86_64.whl
tensorflow-addons==0.15.0
tensorflow-probability==0.15.0
tensorflow-model-optimization
# MILP solvers
cplex==12.9.0.0
docloud==1.0.375
docplex==2.10.155
ortools==9.1.9490
